Refactor: 타입 일치와 Config 설정
user, star, order int => Long 타입
queryDSL config 합체 후 common 디렉토리로 이동

+++ back/tingle/src/main/java/com/example/tingle/store/service/impl/OrderServiceImpl.java
@@ -45,7 +45,7 @@ public class OrderServiceImpl implements OrderService {
 
 
     @Override
-    public boolean processOrder(Long productId, Integer fanId, String starName) {
+    public boolean processOrder(Long productId, Long fanId, String starName) {
         if (decreaseProductStock(productId)) {
             saveOrderInformation(productId, fanId);
             return true;
@@ -66,7 +66,7 @@ public class OrderServiceImpl implements OrderService {
         return false;
     }
 
-    private void saveOrderInformation(Long productId, Integer fanId) {
+    private void saveOrderInformation(Long productId, Long fanId) {
         // 주문 정보 저장 로직을 구현
         // 예를 들어, 주문 정보를 데이터베이스에 저장하거나 다른 서비스로 전송할 수 있습니다.
         // 여기서는 단순히 주문 정보를 OrderRepository를 통해 저장하는 예시를 보여줍니다.
